requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
opencv-python>=4.9.0  # NumPy 2.x対応版
Pillow>=10.0.0

# OCR
pytesseract==0.3.10

# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Environment variables
python-dotenv==1.0.0
